:recycle: update and optimize the code

+++ part-4/section-27/the-wild-oasis/src/services/apiCabins.js
@@ -11,64 +11,58 @@ export async function getCabins() {
   return data;
 }
 
-export async function createCabin(newCabin) {
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     '/',
     ''
   );
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  // 1. Create/edit cabin
+  let query = supabase.from('cabins');
 
-  // 1. Create cabin
-  const { data, error } = await supabase
-    .from('cabins')
-    .insert([{ ...newCabin, image: imagePath }])
-    .select();
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq('id', id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
-    throw new Error('Cabins could not be created');
+    throw new Error('Cabin could not be created');
   }
 
   // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from('cabin-images')
     .upload(imageName, newCabin.image);
 
-  // 3. Delete cabin IF there was an error uploading image
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from('cabins').delete().eq('id', data.id);
-
     console.error(storageError);
     throw new Error(
-      'Cabin image could not be uploaded, and therefore the cabin was not created'
+      'Cabin image could not be uploaded and the cabin was not created'
     );
   }
 
   return data;
 }
 
-// export async function updateCabin() {
-//   const { data, error } = await supabase
-//     .from('cabins')
-//     .update({ other_column: 'otherValue' })
-//     .eq('some_column', 'someValue')
-//     .select();
-
-//   if (error) {
-//     console.error(error);
-//     throw new Error('Cabins could not be deleted');
-//   }
-
-//   return data;
-// }
-
 export async function deleteCabin(id) {
   const { data, error } = await supabase.from('cabins').delete().eq('id', id);
 
   if (error) {
     console.error(error);
-    throw new Error('Cabins could not be deleted');
+    throw new Error('Cabin could not be deleted');
   }
 
   return data;
